+++ Mods ChoGGi/RC Bulldozer/metadata.lua
@@ -1,7 +1,7 @@
 return PlaceObj("ModDef", {
-  "title", "RC Bulldozer v0.2",
-  "version", 2,
-  "saved", 1539518400,
+  "title", "RC Bulldozer v0.4",
+  "version", 4,
+  "saved", 1540036800,
 	"image", "Preview.png",
   "tags", "Buildings",
   "id", "ChoGGi_RCBulldozer",
@@ -12,12 +12,11 @@ return PlaceObj("ModDef", {
 	},
 	"lua_revision", LuaRevision,
   "description", [[Flattens the ground in front of it.
-This bulldozer is very shy when it comes to bumpy ground...
 
-Options to change the radius, show a circle where it flattens, and change the ground texture.
+Options to change the radius, show a circle where it flattens, and change the ground texture (or turn it off).
+The buildable area won't be updated till you turn off Dozer Toggle (expensive function call).
+
 
 Known issues:
-The sides of ground don't look that good, you can use Terraformer to smooth them.
-It has to re-build the buildable terrain grid whenever it stops, so slight lag spike.
-You have to fiddle with getting it to drive in the right spot (like I said shy), drop/remove a construction site to update driveable.]],
+The sides of ground don't look that good, you can use Terraformer to smooth them.]],
 })